Set Default TicTacToe
The default game file. Realistically, I shouldn't have written anything in the player class but we should really only need those three things.

When implimenting Gene, we should try to create an interface for players so we cab separate human and a.i. but I wasn't sure how to deal with that in python